refactor(models): move model definitions into models package (Closes #52)
